Chore/rhc templates cleanup (#612)
