+++ apps/rhc-templates/src/app/layout.tsx
@@ -4,9 +4,10 @@ import { PropsWithChildren } from 'react';
 import '@nl-rvo/assets/fonts/index.css';
 import '@rijkshuisstijl-community/design-tokens/dist/index.css';
 import '@rijkshuisstijl-community/components-css/index.scss';
+import { Logo } from '@rijkshuisstijl-community/components-react';
+import { NavBar, NavList, NavListLinkButton, PageHeader } from '@utrecht/component-library-react/dist/css-module';
 import '@rijkshuisstijl-community/font/src/index.mjs';
 import './globals.css';
-import './page/page.scss';
 import Link from 'next/link';
 import Head from 'next/head';
 
@@ -19,25 +20,36 @@ export default function RootLayout({ children }: PropsWithChildren<{}>) {
         <title>Rijkshuisstijl demo - Index</title>
       </Head>
       <body>
-        <div className="information">
-          <div className="unstarted">Elemententen die nog helemaal geen styling hebben zijn met rood omgeven</div>
-          <div className="unfinished">
-            Elementen die al styling hebben maar die nog niet helemaal af zijn zijn met geel omgeven
-          </div>
-          <div>
-            Elementen die niet met een kleur zijn omgeven zijn af, let op: dit betekent niet per se dat ze ook in
-            storybook staan.
-          </div>
-          <Link href="/page">Page</Link>
-          <br />
-          <Link href="/form">Form</Link>
-          <br />
-          <Link href="/details">Details</Link>
-          <br />
-          <Link href="/collage">Collage</Link>
-        </div>
         <RHCTheme>
-          <Document>{children}</Document>
+          <main className="page">
+            <div className="rhc-background-color-white">
+              <PageHeader>
+                <div className="unfinished">
+                  <Logo organisation="Rijkshuisstijl Community"></Logo>
+                </div>
+              </PageHeader>
+            </div>
+            <div className="unstarted rhc-background-color-lint-blauw">
+              {/* wrong spacing, wrong background color, wrong text color, no way to add right aligning link */}
+              <NavBar>
+                <NavList>
+                  <NavListLinkButton>
+                    <Link href="/">Page</Link>
+                  </NavListLinkButton>
+                  <NavListLinkButton>
+                    <Link href="/form">Form</Link>
+                  </NavListLinkButton>
+                  <NavListLinkButton>
+                    <Link href="/details">Details</Link>
+                  </NavListLinkButton>
+                  <NavListLinkButton>
+                    <Link href="/collage">Collage</Link>
+                  </NavListLinkButton>
+                </NavList>
+              </NavBar>
+            </div>
+            <Document>{children}</Document>
+          </main>
         </RHCTheme>
       </body>
     </html>
