--- apps/rhc-templates/src/app/page/page.scss
@@ -1,41 +1,0 @@
-.rhc-background-color-white {
-  background-color: var(--rhc-color-wit);
-}
-
-.rhc-background-color-lint-blauw {
-  background-color: var(--rhc-color-lintblauw-500);
-}
-
-.rhc-spacing-2 {
-  margin-block-end: var(--rhc-space-400);
-}
-
-.rhc-main-content {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.utrecht-figure {
-  margin-inline-start: 0;
-  margin-inline-end: 0;
-}
-
-.rhc-page-content {
-  --utrecht-page-content-padding-block-end: var(--rhc-space-600);
-  --utrecht-page-content-padding-block-start: var(--rhc-space-600);
-  --utrecht-page-padding-inline-end: var(--rhc-space-700);
-  --utrecht-page-padding-inline-start: var(--rhc-space-700);
-
-  margin-block-start: 80px;
-  margin-block-end: 80px;
-  width: 70%;
-}
-
-@media (max-width: 780px) {
-  .rhc-page-content {
-    width: 100%;
-    margin-block-start: 0;
-    margin-block-end: 0;
-  }
-}
